backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase==2.4.3
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
pydantic[email]>=2.9.0
python-multipart>=0.0.9
httpx>=0.27.0
